+++ packages/wonder-blocks-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @khanacademy/wonder-blocks-core
 
+## 4.3.1
+
 ## 4.3.0
 
 ### Minor Changes